encapsulate SslHttpConfigurator
